Cruds das entidades p1

+++ database/migrations/2020_07_26_190947_create_contatos_table.php
@@ -16,7 +16,7 @@ class CreateContatosTable extends Migration
         Schema::create('contatos', function (Blueprint $table) {
             $table->bigIncrements("id");
             $table->string("nome");
-            $table->integer("telefone");
+            $table->bigInteger("telefone");
             $table->string("email",100)->unique();
             $table->softDeletes();
             $table->timestamps();
